Improve temporary zip file handling
Move mktemp() closer to the zip creation to tighten the window for
the race condition. This should still be replaced with a safe solution,
see source comment.

Never try to remove the temporary file if it already exists, that's silly
and dangerous (it probably belongs to a different process); just give up.
Update the error message for this case.

Fix an error path where the the code attempts to remove the template
pattern instead of the temporary file.