Don't follow symlinks, only act on regular files
Following symlinks is problematic for directories since there may
be cycles leading to endless recursion; replacing a symlink to a
zip file with a new zip file isn't what you usually want either.

Trying to process any special file (device, named pie or socket)
seems like a bad idea.

Note that it would be desirable to skip these files only during
recursion and handle everything explicitly listed on the command
line anyway (symlinks, special files, "." and "..", files not
matching *.zip) because the user is always right and may have a
use case for that. But that would require some refactoring and
isn't implemented in this commit.