Log file fixes
Move closing process logs out of DisplayMigrateSummary() since it it's
never called on critical errors. This avoid leaking open log files in
these cases.

The error log file is currently always created in the current directory,
so we mustn't look for it in pszStartPath (only different on WIN32).

Try harder not to remove error logs which contain messages. Note that
this isn't possible to handle correctly without a design change in case
several trrntzip instances run concurrently. Add a comment which explains
the issue.