Don't attempt to change slashes in zip entries
A backslash in a zip entry isn't a bad path separator but a perfectly
valid ordinary character in a file name on many systems (despite the
bad wording in the zip appnote and the fact that it can't be used on
Windows), so it shouldn't be changed.

Moreover, it doesn't work.

For example, a file back\slash.txt in back\slash.zip yields the following
error:
  Rezipping - back\slash.zip
  --------------------------------------------------
  Unable to open "back/slash.txt" from "back\slash.zip"
  Execution time 0 hours 0 mins 0 secs
(The \ in the zip name doesn't matter, changing \ to / in the entry
breaks things.)

Just delete the bad code.