Fix .zip extension check
EndsWithCaseInsensitive() is symmetric, i.e., it checks whether the
shorter string appears at the end of the longer one, the order doesn't
matter. However that isn't what we want when checking for a file
extension; e.g. a file named "p" would appear to end with ".zip" while
it's the other way around.

Fix the check by rewiting EndsWithCaseInsensitive() such that it always
checks whether the string in the second parameter appers at the end of
the first. Also return nonzero (true) on success, as inverted logic is
confusing for predicate functions.