New utility function to grow a dynamic string array geometrically
The places where this was implemented before did different checks
for corner cases and neither was entirely correct.

Create a new function DynamicStringArrayGrow() for this task it
and use it, so there's only a single place to update if we want to
tweak the behaviour in the future.

Update DynamicStringArrayResize() such that in can handle an new
size of zero and destroy the array in this case. This simplifies
the implementation of the former function.