Check for more errors while reading zip entries
A corrupted central directory entry or a name we can't handle could
result in a truncated file list and then all remaining entries would
be silently dropped.

Add some checks to GetFileList() to prevent data loss in such situations,
and refuse to process affected zip files. Change the signature of the
function to return an error code instead of the file list, such that
we can distinguish between allocation errors and zip file errors. The
file list is updated directly in the workspace, which replaces file
name array and length parameters.