Decrease buffer size for re-compression
A 4 MiByte buffer to hold decompressed data that gets immediately
re-compressed seems a bit excessive. A few KiBytes should be enough
to hide the overhead of calling the unzip and zip functions repeatedly,
larger buffers could even reduce cache efficiency.

In practice, a test with different buffer sizes shows that the changes
in performance on the test machine are well below the noise level for
sizes in the range between 8 KiByte and 8 MiByte (i.e., the effect on
the cache seems negligible here), while performance drops significantly
with 1 KiByte and below (which is expected).

Let's pick 64 KiByte. That's well above the threshold where the tiny
buffer overhead vanishes, but can still be considered small.