Open error log in the same directory as the other logs
The process log files are created in the "startup path", which can
differ from the current directory on WIN32. Create error.log in the
same directory and store the full path name in WORKSPACE.

While at it, ensure a "There were problems!..." message is printed
to the console at the very end if appropriate, not when it's written
into the process log, and also consider the case that no error log
exists.