Clarify license, move license text out of COPYING, update GPL2
State more clearly that the program's license is GPL-2 or later.
Remove the actual license texts from COPYING, and add an up-to-date
copy of GPL-2 in a separate file. The full text of the minizip/zlib
license is already included in every relevant file, no need to
repeat it here.