Update minizip source to version from zlib-1.3.1
Import unmodified copies of the MiniZip files from zlib-1.3.1,
copy the internals needed by TorrentZip to a new header file
src/minizip.h (as opposed to modifying the MiniZip files), and
reinstate the two DISABLED_FOR_TORRENTZIP exclusions.

Adapt trrntzip.c to the changes.

This update finally drops the old style declarations from the
MiniZip code, and the local modifications are limited to two
disabled code blocks and a comment in zip.c.

Note that this drops the modifications in ioapi.h (which I assume
are unnecessary; should be fixed upstream otherwise) and it
effectively reverts commit c3faf1370b3a ("Fix old bug in zip64
support.") which was rejected upstream.