Fix broken detection of missing or wrong TORRENTZIPPED-... comment
The commit listed below introduces a regression where in MigrateZip(),
variable rc would be clobbered when its value is still needed. That
completely breaks handling of zips that should be re-zipped because
of missing or out of date TORRENTZIPPED-... comment.

Rewrite the offending code such that it no longer touches the variable.

Fixes: cf11c900079e ("Check for more errors while reading zip entries")