Add a pair of working tests for zip file entries with long names
Trrntzip currently supports entries up to the length of 1023 bytes.
A zip file with an entry of this length shall be processed correctly,
and a zip file with en entry exeeding this length (by 1 byte in the
test case) shall be rejected.

These tests will need to be revisited when the implementation limit
gets lifted.