Add a pair of tests for .zip extension check
TorrentZip should consider any file given explicitly as command line
argument, regardless of extension and type. On the other hand, when
processing a directory, only regular files with a .zip extension (case
insensitive) and subdirectories should be considered.

Add tests to check that files with different extensions are processed
or ignored as expected. We can't test with symlinks, sockets, or fifos
since nihtest doesn't support their creation in the sandbox.