Run vcpkg directly
The lukka/run-vcpkg action doesn't accept the inputs it's used with,
and it fails to install zlib. It seems it needs a vcpkg.json file.
For our purpose, the action looks overly complicated. Just running
"vcpkg install" should be good enough for now. Binary caching can
be considered later if desired.