Mark tests failing on Windows
Four test fail because the directory separator is different on
Windows; the tests need to be fixed to accept either variant.

In case of directories-2-d.test, the "Directory ${dir}/ Removed"
messages may appear in any order (it depends on the implementation
details of qsort()). Though the test happens to work on the other
major systems, it isn't valid anywhere in its current state and
needs to be fixed or disabled.