Make sorting by basename fully deterministic
BasenameCompare() completely ignored leading directories.
While this is fine with respect to the zip files trrntzip creates,
the order of printed messages can differ when directory entries
get removed or re-zipping fails because of name clashes.

Compare the full paths of entries in case the basenames compare
equal to always ensure the same sort order and thus deterministic
output.

This should fix the flaky directories-2-d.test, so remove it from
XFAIL_TESTS on WIN32.